Improve code for day 3
In an effort to make the code tell a better story, I did the following:
- moved the main functionality into a function named after the
  terminology used
- added test code for the example
- detected the bit width, instead of using a const

Detecting the bit width adds a little bit of complexity, but I think it
makes it easier to understand than passing a width variable around. This
is mainly to handle the fact that the example uses a different width
than the "real" input.